Add FUZZING_EMBED_DATA option to embed secp256k1 precomputed data
In this old version of system scripts, a 1 MB of secp256k1
precomputation table is required for the scripts. It really does not
make much sense to tweak this 1 MB data in the fuzzing engine(since the
fuzzing engine will spend the majority of time tweeaking those data, not
the actual transaction data we care). This commit addsa new mode where
one can embed the data as part of the program, so fuzzing engines can
focus on real important data.

That being said, the newly added mode is optional, one is perfectly fine
treating those additional data as part of input, and let the fuzzing
engine tweak them for different behaviors.

When turning on the newly added mode(secp256k1 data are removed from
fuzzing corpus data and treated as part of program), code coverage
increases at a much faster speed than before.